Some final cleanups for the FreeBSD cookie code. Namely removing timestamp stuff that doesn't need to be in there.
There is still some commented out code that could be removed and a few WTFs in the way of a possible double-free and locking problem.

For some reason I get panics if I allocate reply data on the stack vs dynamically, see the comments for more details. There's nothing wrong with the dynamic method, though.

Anyway, this all seems to work for me.